refactor(web): extract the keyboard surface into useEditorKeyboard
The shortcut dispatch (runShortcut over shortcuts.ts's catalog, with
selectAllNodes and toggleSelectionLock) and the three keydown handlers the
JSX wires — the canvas root's, the focused resize handle's, and the connect
handle's — move into use-editor-keyboard.ts with their four keyboard
constants, verbatim. The hook consumes the editor's state and appliers and
owns no state of its own, so the editor-state-surface ledger is untouched;
the keyboard resize path's direct onChange fold moves with it, so the file
joins purity-guard's MUTATING_SOURCES command-boundary tripwire.

The extraction forced the real declaration order the inline version hid:
the inputs object reads stepZoom/frameContent/frameSelection as values, so
the hook call sits below their definitions instead of relying on
deferred-call forward references.

SpatialEditor.tsx: 3016 -> 2742 lines. Mutation-checked: disabling the
select-all arm in the hook fails 3 select-all-nudge browser tests; restore
is green. Keyboard browser suites (select-all-nudge, edge-select-delete,
edit-handle, escape-cancels-new-note, duplicate, zoom-to-fit, node-lock,
edge-lock, multi-select, hand-pan-after-gesture, multi-resize — 64 tests)
and the spatial-editor jsdom suite (45 files) pass unchanged.
